+++ tests/test_lab_dev/test_states/test_states_visualization.py
@@ -1,0 +1,96 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for the state visualization."""
+
+# pylint: disable=protected-access, unspecified-encoding, missing-function-docstring, expression-not-assigned, pointless-statement
+
+import json
+from pathlib import Path
+import pytest
+
+
+from mrmustard import math
+from mrmustard.lab_dev.states import Coherent
+
+
+class TestVisualization:
+    r"""
+    Tests the functions to visualize states.
+    """
+
+    # set to ``True`` to regenerate the assets
+    regenerate_assets = False
+
+    # path
+    path = Path(__file__).parent.parent / "assets"
+
+    def test_visualize_2d(self):
+        st = Coherent([0], y=1) + Coherent([0], y=-1)
+        fig = st.visualize_2d(resolution=20, xbounds=(-3, 3), pbounds=(-4, 4), return_fig=True)
+        data = fig.to_dict()
+
+        if self.regenerate_assets:
+            fig.write_json(self.path / "visualize_2d.json", remove_uids=True)
+
+        with open(self.path / "visualize_2d.json") as file:
+            ref_data = json.load(file)
+
+        assert math.allclose(data["data"][0]["x"], ref_data["data"][0]["x"])
+        assert math.allclose(data["data"][0]["y"], ref_data["data"][0]["y"])
+        assert math.allclose(data["data"][0]["z"], ref_data["data"][0]["z"])
+        assert math.allclose(data["data"][1]["x"], ref_data["data"][1]["x"])
+        assert math.allclose(data["data"][1]["y"], ref_data["data"][1]["y"])
+        assert math.allclose(data["data"][2]["x"], ref_data["data"][2]["x"])
+        assert math.allclose(data["data"][2]["y"], ref_data["data"][2]["y"])
+
+    def test_visualize_2d_error(self):
+        with pytest.raises(ValueError):
+            Coherent([0, 1]).visualize_2d(20)
+
+    def test_visualize_3d(self):
+        st = Coherent([0], y=1) + Coherent([0], y=-1)
+        fig = st.visualize_3d(resolution=20, xbounds=(-3, 3), pbounds=(-4, 4), return_fig=True)
+        data = fig.to_dict()
+
+        if self.regenerate_assets:
+            fig.write_json(self.path / "visualize_3d.json", remove_uids=True)
+
+        with open(self.path / "visualize_3d.json") as file:
+            ref_data = json.load(file)
+
+        assert math.allclose(data["data"][0]["x"], ref_data["data"][0]["x"])
+        assert math.allclose(data["data"][0]["y"], ref_data["data"][0]["y"])
+        assert math.allclose(data["data"][0]["z"], ref_data["data"][0]["z"])
+
+    def test_visualize_3d_error(self):
+        with pytest.raises(ValueError):
+            Coherent([0, 1]).visualize_3d(20)
+
+    def test_visualize_dm(self):
+        st = Coherent([0], y=1) + Coherent([0], y=-1)
+        st.manual_shape[0] = 20
+        fig = st.visualize_dm(20, return_fig=True)
+        data = fig.to_dict()
+
+        if self.regenerate_assets:
+            fig.write_json(self.path / "visualize_dm.json", remove_uids=True)
+
+        with open(self.path / "visualize_dm.json") as file:
+            ref_data = json.load(file)
+        assert math.allclose(data["data"][0]["z"], ref_data["data"][0]["z"])
+
+    def test_visualize_dm_error(self):
+        with pytest.raises(ValueError):
+            Coherent([0, 1]).visualize_dm(20)
